fixed bug

+++ lab5/src/LocalNavigation/FSM.java
@@ -183,7 +183,7 @@ public class FSM {
 		}
 	}
 	
-	private static final float POST_RETREAT_ROTATION_SPEED = 0.4;
+	private static final float POST_RETREAT_ROTATION_SPEED = (float) 0.4;
 
 	// If we're aligned, we move away from the obstacle
 	//
